% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_synthetic_compare.R
\name{compare_real_synthetic}
\alias{compare_real_synthetic}
\title{Compare real and synthetic data}
\usage{
compare_real_synthetic()
}
\arguments{
\item{df}{mandatory, dataframe \cr
a dataframe with APPID and STUDYID columns}
}
\value{
App
}
\description{
This is an shiny app that can be used to visually inspect
real and synthetic data
}
\examples{
\dontrun{
compare_real_synthetic()
}
}
