% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_synthetic_compare.R
\name{compare_real_synthetic_shiny}
\alias{compare_real_synthetic_shiny}
\title{Compare real and synthetic data}
\usage{
compare_real_synthetic_shiny()
}
\value{
App
}
\description{
This is an shiny app that can be used to visually inspect
real and synthetic data
}
\examples{
\dontrun{
compare_real_synthetic()
}
}
