% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_to_func.R
\name{sanitize}
\alias{sanitize}
\title{function to generate fake data}
\usage{
sanitize(path, number = 1, recovery = FALSE, where_to_save = NULL)
}
\arguments{
\item{path}{Mandatory\cr
path where real data/xpt files located, should be a directory that contains
xpt files}

\item{number}{mandatory, default 1\cr
how many studies to generate. Currently only work with 1}

\item{recovery}{optional\cr
recovery}

\item{where_to_save}{mandatory\cr
where to save generated xpt files. Should be a directory.}
}
\description{
function to generate fake data
}
