% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data_app.R
\name{sanitize_shiny_app}
\alias{sanitize_shiny_app}
\title{Generate synthetic data from real data}
\usage{
sanitize_shiny_app()
}
\value{
app
}
\description{
Generate synthetic data with shiny app
}
\examples{
\dontrun{
sanitize_shiny_app()
}
}
