% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_to_func.R
\name{sanitize}
\alias{sanitize}
\title{Run function}
\usage{
sanitize(path, number = 1, recovery = FALSE, where_to_save = NULL)
}
\arguments{
\item{path}{Mandatory\cr
path where real data located, should be directory}

\item{number}{mandatory, default 1\cr
how many studies to generate}

\item{recovery}{optional\cr
recovery}

\item{where_to_save}{optional\cr
if no directory path given, fake data will be save
directory given in path argument}
}
\description{
Run function
}
